se cambio el precio a double, con todos los tests. se reescribio el constructor de compra

+++ libreria/src/main/java/ar/edu/untref/aydoo/ArticuloDeLibreria.java
@@ -6,12 +6,12 @@ public class ArticuloDeLibreria extends Producto{
         super(precio);
     }
 
-    public int getPrecio(){
+    public double getPrecio(){
        return precioConIva();
     }
 
-    public int precioConIva(){
-        return (int)(this.precio*1.21);
+    public double precioConIva(){
+        return (this.precio*1.21);
     }
 
 }
