+++ libreria/src/main/java/ar/edu/untref/aydoo/PublicacionRegular.java
@@ -4,7 +4,7 @@ public class PublicacionRegular extends Producto{
 
     private int periodicidadMensual;
 
-    public PublicacionRegular(int precio, int periodicidadPorMes){
+    public PublicacionRegular(double precio, int periodicidadPorMes){
         super(precio);
         this.periodicidadMensual = periodicidadPorMes;
     }
